feat: custom handle of errors

+++ src/storage/groups/createGroup.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { AppError } from '@utils/AppError';
 
 import { GROUP_COLLECTION } from '../storageConfig';
 import { getAllGroups } from './getAllGroups';
@@ -7,6 +8,13 @@ import { getAllGroups } from './getAllGroups';
 export async function createGroup(newGroup: string) {
   try {
     const storedGroups = await getAllGroups();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError('Esse grupo já existe.');
+    }
+
     const storage = JSON.stringify([...storedGroups, newGroup]);
 
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
